semana02/dia6/requirements.txt
#FLASK : Microframework web
Flask
#FLASK-SQLALCHEMY : ORM para manipular bases de datos relacionales
Flask-SQLAlchemy
#flask-marshmallow : sirve para crear modelos del orm y serializarlos
flask-marshmallow
#marshmallow-sqlalchemy : integra marshmallow con sqlalchemy
marshmallow-sqlalchemy
#mysqlclient : libreria para conectar mysql con python
mysqlclient
#gunicorn : servidor web para python en heroku
gunicorn
#flask-cors : para permitir acceso de otros servidores
flask-cors
#python-dotenv : para variables de entorno en python
python-dotenv